feat: CU-869877xmz - Add healthcheck endpoint
